Add some padding in rxConfig_t to compensate for size reduction
Since MAX_MAPPABLE_RX_INPUTS was reduced from 8 to 4, restoring the
PG when upgrading would corrupt several fields in rxConfig_t. This
introduces 4 bytes of padding just after it as well as a couple of
notes to remind whoever increases the PG number, should also delete
the padding and the TODO. This way users won't have to manually
restore the RX configuration when upgrading.